+++ src/circuit/freeze.rs
@@ -9,7 +9,7 @@
 // FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
 // details. You should have received a copy of the GNU General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-//! Circuit for the freezing/releasing of auditable anonymously-transferable
+//! Circuit for the freezing/releasing of user configurable
 //! asset.
 use crate::{
     circuit::{